+++ app/forms/conversion/task/funding_agreement_contact_task_form.rb
@@ -1,34 +1,20 @@
 class Conversion::Task::FundingAgreementContactTaskForm < ::BaseTaskForm
-  attribute :name, :string
-  attribute :email, :string
-  attribute :title, :string
-
-  validates :name, :email, :title, presence: true
-  validates :email, format: {with: URI::MailTo::EMAIL_REGEXP}
+  attribute :funding_agreement_contact_id, :string
 
   def initialize(tasks_data, user)
     @tasks_data = tasks_data
     @user = user
-    @project = tasks_data.project
-    @contact = Contact::Project.find_or_create_by(project: @project, funding_agreement_contact: true)
+    @project = @tasks_data.project
 
     super(@tasks_data, @user)
-
-    assign_attributes(
-      name: @contact.name,
-      title: @contact.title,
-      email: @contact.email
-    )
+    self.funding_agreement_contact_id = @project.funding_agreement_contact_id
   end
 
   def save
-    @contact.assign_attributes(
-      project: @project,
-      name: name,
-      title: title,
-      email: email,
-      funding_agreement_contact: true
-    )
-    @contact.save!
+    @project.update!(funding_agreement_contact_id: funding_agreement_contact_id)
+  end
+
+  private def completed?
+    @project.funding_agreement_contact_id.present?
   end
 end